--- MartinWebApp/Views/People/PeopleSearch.cshtml
@@ -1,48 +1,0 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
-@model MartinWebApp.Models.PeopleViewModel
-@{
-    ViewData["Title"] = "PeopleIndex";
-}
-<h1>People Index</h1>
-<h2>Search People</h2>
-<form asp-controller="People" asp-action="PeopleSearch" method="post">
-    <label>Name:</label>
-    <input type="text" name="name" />
-    <label>Tele:</label>
-    <input type="text" name="phonenumber" />
-    <label>City:</label>
-    <input type="text" name="city" />
-
-    <input type="submit" value="Search" />
-</form>
-
-<h2>Add Person</h2>
-<form asp-controller="People" asp-action="PeopleIndex" method="post">
-    <label>Name:</label>
-    <input type="text" name="name" />
-    <label>Tele:</label>
-    <input type="text" name="phonenumber" />
-    <label>City:</label>
-    <input type="text" name="city" />
-
-    <input type="submit" value="Add" />
-</form>
-@if (ViewBag.ErrorMess != null)
-{
-<p>@ViewBag.ErrorMess</p>
-
-}
-<h2>Search Results</h2>
-<table>
-    @foreach (var p in MartinWebApp.Models.PeopleViewModel.peopleResult)
-    {
-    <tr>
-        <td>Name: @p.Name</td>
-        <td>Tele: @p.PhoneNumber</td>
-        <td>City: @p.City</td>
-    </tr>
-
-    }
-</table>
