+++ MartinWebApp/Views/Shared/_Layout.cshtml
@@ -19,7 +19,7 @@
                     <li><a asp-area="" asp-controller="Home" asp-action="Projects">Projects</a></li>
                     <li><a asp-area="" asp-controller="Fever" asp-action="FeverCheck">Fever</a></li>
                     <li><a asp-area="" asp-controller="Guessing" asp-action="GuessingGame">GuessingGame</a></li>
-                    <li><a asp-area="" asp-controller="People" asp-action="">People</a></li>
+                    <li><a asp-area="" asp-controller="Home" asp-action="People">People</a></li>
                 </ul>
             </nav>
         </header>
